% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_features}
\alias{add_features}
\title{Add features}
\usage{
add_features(mts_data, xreg_data = NULL)
}
\arguments{
\item{mts_data}{A univariate or multivariate 'ts', 'mts' or 'xts' object}

\item{xreg_data}{A univariate or multivariate 'ts', 'mts' or 'xts' object, optional external regressors}
}
\value{
A univariate or multivariate 'xts' object
}
\description{
Add features
}
