% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_timesteps}
\alias{add_timesteps}
\title{Add timesteps
Description:
for lstm, when the number of time steps (i.e. number of lags) is defined,
for each input variable, the required number of lagged values needs to be
added as inputs to the input matrix}
\usage{
add_timesteps(
  data_df,
  fc_horizon = 12,
  valid_set_size = 12,
  tsteps = 12,
  lag_setting = 12,
  backtesting_opt = list(use_backtesting = FALSE, backtesting_nb_iter = 1,
    backtesting_method = c("rolling", "moving"), backtesting_set_size = c("expanding",
    "fixed"))
)
}
\value{

}
\description{
when the ts is seasonal with periodicity = 12, then lag_setting should be 12.
The volume of the same month of previous year will most likely have a high
influence on today's volume.
when the ts is non-seasonal, lagSetting = 1, i.e. the volume of previous
month will most likely have the strongest impact on today.
}
\details{
when lag_setting = 12 and tsteps < lagSetting (e.g. tsteps=4),
then input values will be: t-12, t-11, t-10, t-9
when lag_setting = 12 and tsteps = lagSetting, then input values will be:
t-12, t-11, ... , t-1
when lagSetting = 12 and tsteps > lagSetting (e.g. tsteps = 15),
then input values will be: t-15, ... , t-12, ... , t-1

Same goes for lagSetting other than 12.
}
