% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_analyze.R
\name{analyze_fc}
\alias{analyze_fc}
\title{Analyze forecasts}
\usage{
analyze_fc(
  fc,
  ts_name = NULL,
  period = NULL,
  model_names = NULL,
  hybrid_models = NULL,
  error_measure = c("mse, mae, mape"),
  Plot = TRUE
)
}
\arguments{
\item{fc}{A list}

\item{ts_name}{A string}

\item{period}{A numeric}

\item{model_names}{A vector of strings}

\item{hybrid_models}{A list}

\item{error_measure}{A vector of strings}

\item{Plot}{A boolean}
}
\value{
A data frame
}
\description{
Analyze forecasts
}
