% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_backtesting_opt}
\alias{check_backtesting_opt}
\title{Check backtesting options
This function ensures that the selected backtesting options are valid.
If not, the function throws an error. If the options are NULL, then default
values will be applied.}
\usage{
check_backtesting_opt(backtesting_opt)
}
\arguments{
\item{backtesting_opt}{NULL or a list}
}
\value{
A list of backtesting options
}
\description{
Check backtesting options
This function ensures that the selected backtesting options are valid.
If not, the function throws an error. If the options are NULL, then default
values will be applied.
}
