% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_colnames}
\alias{check_colnames}
\title{Drop punctuations in colnames}
\usage{
check_colnames(data_colnames, default_colname = "time_series", nb_colnames = 1)
}
\arguments{
\item{default_colname}{a string,}

\item{colnames_vec}{a vector of strings, colnames to be cleaned}
}
\value{
colnames_vec
}
\description{
Punctuations in colnames will dropped in order to
avoid errors when converting between different data types.
}
