% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_data_dir}
\alias{check_data_dir}
\title{Check the filepath where forecasts can be saved}
\usage{
check_data_dir(data_dir)
}
\arguments{
\item{data_dir}{A string, directory to which results can be saved as text files}
}
\value{
NULL or a valid filepath
}
\description{
This function ensures that the filepath where forecast can be saved is valid
If not the argument is invalid, the function throws an error.
}
