% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_data_sv_as_xts}
\alias{check_data_sv_as_xts}
\title{Check the time series data and convert to xts}
\usage{
check_data_sv_as_xts(input_data, default_colname = "time_series")
}
\arguments{
\item{input_data}{ts, mts or xts object}
}
\value{
xts object with punctuations dropped in colnames
}
\description{
This function ensures that the input data is of type xts, ts or mts and
converts it to an xts object. Furthermore, if no colnames are specified, default
colnames will be applied. Moreover, punctuations in colnames will dropped to
avoid errors when converting between different data types.
}
