% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fc_horizon}
\alias{check_fc_horizon}
\title{Check forecasting horizon
This function ensures that the selected forecasting horizon is valid.
If not, the function throws an error.}
\usage{
check_fc_horizon(fc_horizon)
}
\arguments{
\item{fc_horizon}{must be a positive integer}
}
\value{
fc_horizon: a positive integer
}
\description{
Check forecasting horizon
This function ensures that the selected forecasting horizon is valid.
If not, the function throws an error.
}
