% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_models_args}
\alias{check_models_args}
\title{Check models' arguments}
\usage{
check_models_args(models_args, model_names = NULL)
}
\arguments{
\item{models_args}{A list, optional arguments to passed to the models}

\item{model_names}{A list or vector of strings representing the model names to be used}
}
\value{
models_args: A list, parameters to be passed to the models where unused parameters are dropped
}
\description{
This function ensures that the models' arguments are a list. Furthermore, the
function checks if these arguments match the selected model names, otherwise they
will be dropped.
}
