% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_nb_cores}
\alias{check_nb_cores}
\title{Check the number of selected CPU cores}
\usage{
check_nb_cores(nb_cores)
}
\arguments{
\item{nb_cores}{An integer, the number of CPU cores to select}
}
\value{
An integer, the number of CPU cores to select
}
\description{
This function ensures that the user selects a valid number of CPU cores.
If the number of cores is invalid, then use the total number of available cores as default.
}
