% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_period_iter}
\alias{check_period_iter}
\title{Check the period identifier}
\usage{
check_period_iter(period_iter)
}
\arguments{
\item{period_iter}{A string, period identifier of format: 'period' + '_' + iter}
}
\value{
A POSIXct, time identifier
}
\description{
This function ensures that the user specifies a valid period identifier.
}
