% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_preprocess_fct}
\alias{check_preprocess_fct}
\title{Check the custom preprocessing function}
\usage{
check_preprocess_fct(fct)
}
\arguments{
\item{fct}{A function, the custom preprocessing function to handle missing values in the data}
}
\value{
NULL (if no fct is specified) or fct (if fct is specified)
}
\description{
This function checks whether the specified function is indeed a function
}
