% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_save_fc_to_file}
\alias{check_save_fc_to_file}
\title{Check the filepath where forecasts can be saved}
\usage{
check_save_fc_to_file(save_fc_to_file)
}
\arguments{
\item{save_fc_to_file}{NULL or a valid filepath}
}
\value{
NULL or a valid filepath
}
\description{
This function ensures that the filepath where forecast can be saved is valid
If not the argument is invalid, the function throws an error.
}
