% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_tmp_test_set_size}
\alias{check_tmp_test_set_size}
\title{Check optional test set size
This function ensures that the optional test set size is valid. The optional
test set size is used to identify the best performing model inside the
h2o.automl procedure. If the argument is invalid, the function throws an error.}
\usage{
check_tmp_test_set_size(tmp_test_set_size)
}
\arguments{
\item{tmp_test_set_size}{must be a positive integer}
}
\value{
tmp_test_set_size: a positive integer
}
\description{
Check optional test set size
This function ensures that the optional test set size is valid. The optional
test set size is used to identify the best performing model inside the
h2o.automl procedure. If the argument is invalid, the function throws an error.
}
