% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_use_parallel}
\alias{check_use_parallel}
\title{Check parallel processing option}
\usage{
check_use_parallel(use_parallel)
}
\arguments{
\item{use_parallel}{A boolean}
}
\value{
A boolean
}
\description{
This function ensures that the user selects a valid parallel processing option.
If use_parallel = True, then time series will be processed in parallel.
}
