% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{collapse_model_par}
\alias{collapse_model_par}
\title{Store model estimates as a string}
\usage{
collapse_model_par(model_par_vec)
}
\arguments{
\item{model_par_vec}{A vector of strings representing the model estimates which will be collapsed to a single string.}
}
\value{
A string
}
\description{
This function collapses a vector to a single string where values are separated by ';'. This conversion
is useful when estimates are later stored in a data table, providing the user the possibility to check on
the most important model parameters.
}
