% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{combine_fc_results}
\alias{combine_fc_results}
\title{Combine forecasting info}
\usage{
combine_fc_results(
  ts_name,
  model_name,
  fc_formated,
  actual_formated,
  split_keys,
  model_descr = NULL,
  model_par = NULL,
  model_args = NULL,
  period_iter = NULL,
  time_id = NULL,
  ...
)
}
\arguments{
\item{ts_name}{A string}

\item{model_name}{A string representing the model name to be used}

\item{fc_formated}{A data.frame}

\item{actual_formated}{A data.frame}

\item{split_keys}{A data.frame, with the sample split keys (e.g. train, valid, test)}

\item{model_descr}{A string}

\item{model_par}{A string representing the model estimates which were previously collapsed into a single string}

\item{model_args}{A list, optional arguments to pass to the models}

\item{period_iter}{A string, period id (format: 'period' + '_' + iter). This id defines the iteration number of forecasting
exercise.}

\item{time_id}{A POSIXct, timestamp created with \code{\link[base]{Sys.time}} which is then appended to the results}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A data.frame object
}
\description{
This function combines every info which will be stored in a data table.
}
