% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{default_prepro_fct}
\alias{default_prepro_fct}
\title{Default preprocessing function}
\usage{
default_prepro_fct(ts_data)
}
\arguments{
\item{ts_data}{A ts or xts object}
}
\description{
Default preprocessing function to handle missing values in the data. This function selects the
largest interval of non-missing values and attributes the most recent dates to these values.
}
