% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{extract_coef_ets}
\alias{extract_coef_ets}
\title{Extract model estimates for ETS}
\usage{
extract_coef_ets(forecast_obj)
}
\arguments{
\item{forecast_obj}{A 'forecast' object}
}
\value{
A vector of strings
}
\description{
This function extracts the estimates of the model parameters
}
