% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{extract_coef_snaive}
\alias{extract_coef_snaive}
\title{Extract model estimates for seasonal naive}
\usage{
extract_coef_snaive(fc_obj)
}
\arguments{
\item{fc_obj}{A forecast object}
}
\value{
A vector of strings
}
\description{
This function extracts the estimates of the model parameters
}
