% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{format_historical_data}
\alias{format_historical_data}
\title{Format original data}
\usage{
format_historical_data(data_xts)
}
\arguments{
\item{data_xts}{An xts object, the original (i.e. unprocessed) time series data}
}
\value{
A data.frame object
}
\description{
This function ensures that the original data is uniformely formatted across all forecasting procedures
before saving it in a data table. The original data can be accessed under 'values' with 'key'=='actual'.
}
