% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_engine_api.R
\name{generate_fc}
\alias{generate_fc}
\title{Forecasting Engine API}
\usage{
generate_fc(
  mts_data,
  fc_horizon = 1,
  xreg_data = NULL,
  backtesting_opt = list(use_backtesting = FALSE, backtesting_nb_iters = 1,
    backtesting_method = c("rolling", "moving"), backtesting_set_size = c("expanding",
    "fixed")),
  model_names = c("arima", "ets", "tbats", "bsts", "stl", "snaive", "nnetar"),
  models_args = NULL,
  save_fc_to_file = NULL,
  use_parallel = FALSE,
  ...
)
}
\arguments{
\item{mts_data}{ts, mts or xts object}
}
\value{
A list of forecasts for each time series
}
\description{
This function enables the user to select different forecasting algorithms ranging from
traditional time series models (i.e. ARIMA, ETS, STL) to machine learning methods (i.e. LSTM, AutoML).
}
