% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_arima}
\alias{generate_fc_arima}
\title{ARIMA Model
This function applies the auto.arima function from the forecast package on
time series data.}
\usage{
generate_fc_arima(
  ts_data_xts,
  fc_horizon = 12,
  xreg_xts = NULL,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  arima_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
ARIMA Model
This function applies the auto.arima function from the forecast package on
time series data.
}
