% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_automl_h2o}
\alias{generate_fc_automl_h2o}
\title{Automated Machine Learning
This function applies the h2o.automl function from the h2o package on
time series data.}
\usage{
generate_fc_automl_h2o(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  automl_h2o_arg = list(max_models = 5, stopping_metric = "MAE"),
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
Automated Machine Learning
This function applies the h2o.automl function from the h2o package on
time series data.
}
