% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_bsts}
\alias{generate_fc_bsts}
\title{Bayesian Structural Time Series Model
This function applies the bsts function from the bsts package on
time series data.}
\usage{
generate_fc_bsts(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  bsts_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
Bayesian Structural Time Series Model
This function applies the bsts function from the bsts package on
time series data.
}
