% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_ets}
\alias{generate_fc_ets}
\title{Exponential Smoothing Model
This function applies the ets function from the forecast package on
time series data.}
\usage{
generate_fc_ets(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  ets_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
Exponential Smoothing Model
This function applies the ets function from the forecast package on
time series data.
}
