% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_lstm_keras}
\alias{generate_fc_lstm_keras}
\title{Long-Short Term Memory Network}
\usage{
generate_fc_lstm_keras(
  ts_data_xts,
  fc_horizon = 12,
  xreg_xts = NULL,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  lstm_keras_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{A univariate ts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{xreg_xts}{A univariate or multivariate ts, mts or xts object, optional external regressors}

\item{backtesting_opt}{A list, options of the backtesting program}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{lstm_keras_arg}{A list, optional arguments to pass to the lstm network}
}
\value{
A list, forecast object for each forecasted period
}
\description{
Function to apply lstm networks (\code{\link[keras]{layer_lstm}}) from the \code{\link[keras]{keras}} package on
time series data.
}
\examples{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc_lstm_keras(AirPassengers,
                             fc_horizon = 12)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc_lstm_keras(AirPassengers,
                             fc_horizon = 12,
                             backtesting_opt = list(use_backtesting = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc_lstm_keras(AirPassengers,
                             fc_horizon = 6,
                             backtesting_opt = list(use_backtesting = TRUE,
                                                    backtesting_nb_iters = 6))
}
