% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_nnetar}
\alias{generate_fc_nnetar}
\title{Neural Network
This function applies the nnetar function from the forecast package on
time series data.}
\usage{
generate_fc_nnetar(
  ts_data_xts,
  fc_horizon = 12,
  xreg_xts = NULL,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  nnetar_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
Neural Network
This function applies the nnetar function from the forecast package on
time series data.
}
