% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_stl}
\alias{generate_fc_stl}
\title{Season-Trend Decomposition with Loess Model
This function applies the stl function from the forecast package on
time series data.}
\usage{
generate_fc_stl(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  stl_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
Season-Trend Decomposition with Loess Model
This function applies the stl function from the forecast package on
time series data.
}
