% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_tbats}
\alias{generate_fc_tbats}
\title{TBATS Model
This function applies the tbats function from the forecast package on
time series data. The tbats function is only applicable on ts objects.}
\usage{
generate_fc_tbats(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  tbats_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
TBATS Model
This function applies the tbats function from the forecast package on
time series data. The tbats function is only applicable on ts objects.
}
