% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{nb_diffs}
\alias{nb_diffs}
\title{Determine the number of differencing to obtain a stationary time series}
\usage{
nb_diffs(ts_data, ...)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A list, training, validation and test sets
}
\description{
Uses \code{\link[forecast]{nsdiffs}} and \code{\link[forecast]{ndiffs}}
to determine the number of differencing to obtain a stationary time series.
}
\examples{
\dontrun{
library(datasets)
nb_diffs(AirPassengers)
}
}
