% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize the data}
\usage{
normalize_data(data_df)
}
\arguments{
\item{data_df}{A data.frame object}
}
\value{
A list, normalized data and scalers
}
\description{
Normalize the data
}
