% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{print_model_name}
\alias{print_model_name}
\title{Print to console the model name currently selected}
\usage{
print_model_name(model_name)
}
\arguments{
\item{model_name}{A string representing the model name to be used}
}
\description{
Print to console the model name currently selected
}
