% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{read_tsForecastR}
\alias{read_tsForecastR}
\title{Recursive function to read results from tsForecastR object}
\usage{
read_tsForecastR(tsfcr_obj)
}
\arguments{
\item{tsfcr_obj}{A tsForecastR object}
}
\value{
A data frame
}
\description{
Recursive function to read results from tsForecastR object
}
