% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{reshape_X}
\alias{reshape_X}
\title{Reshape regressors X}
\usage{
reshape_X(X, tsteps = 1, nb_features = 1)
}
\arguments{
\item{X}{input matrix}

\item{tsteps}{An integer, the number of time steps (i.e. lags) with explanatory power. These will be included as regressors.}

\item{nb_features}{A positive integer, number of features/regressors}
}
\description{
Reshape regressors X
}
