% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{reshape_Y_2d}
\alias{reshape_Y_2d}
\title{Check validation set size
This function ensures that the selected validation set size is valid.
If not, the function throws an error.}
\usage{
reshape_Y_2d(X)
}
\arguments{
\item{valid_set_size}{must be a positive integer}
}
\value{
valid_set_size: a positive integer
}
\description{
Check validation set size
This function ensures that the selected validation set size is valid.
If not, the function throws an error.
}
