% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_as_df}
\alias{save_as_df}
\title{Read forecasts from tsForecastR object}
\usage{
save_as_df(
  tsfcr_obj = NULL,
  data_dir = NULL,
  data_colnames = NULL,
  model_names = NULL,
  ...
)
}
\arguments{
\item{tsfcr_obj}{A 'tsForecastR' object}

\item{data_dir}{A string, directory to which results can be saved as text files.}

\item{data_colnames}{A vector of strings, the names of the time series objects to read the
results from}

\item{model_names}{A vector of strings, the models to read the results from}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A data frame
}
\description{
This function transforms a tsForecastR object into a data.frame object
}
\examples{
\dontrun{
library(datasets)

fc <- generate_fc(AirPassengers)
df <- save_as_df(fc)
print(df)
}
}
