% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_fc_bsts}
\alias{save_fc_bsts}
\title{Save forecasts (for bsts.prediction objects)}
\usage{
save_fc_bsts(
  fc_obj,
  raw_data,
  sample_split,
  save_fc_to_file,
  model_name,
  time_id = base::Sys.time(),
  period_iter = NULL,
  model_args = NULL,
  ...
)
}
\arguments{
\item{fc_obj}{A bsts.prediction object}

\item{raw_data}{A univariate ts or xts object, original (i.e. unprocessed) time series data}

\item{sample_split}{A list, the sample split}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{model_name}{A string, name of the forecasting model}

\item{time_id}{A POSIXct, created with \code{\link[base]{Sys.time}} and appended to results}

\item{period_iter}{A string, period identifier of format: 'period' + '_' + iter}

\item{model_args}{A list, optional arguments to pass to the models}
}
\value{
A data frame
}
\description{
This function extracts the estimates of the model parameters
}
