% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transform_data}
\alias{transform_data}
\title{Apply a transformation on the data}
\usage{
transform_data(
  original_ts,
  transformed_ts = NULL,
  method = "diff",
  apply_transform = TRUE
)
}
\arguments{
\item{original_ts}{ts or xts object, data to apply transformation on}

\item{transformed_ts}{ts or xts object, data on which the transformation has been applied on}

\item{method}{string, method of data transformation}

\item{apply_transform}{boolean, if TRUE the data will be transformed, else the transformation will be undone.}
}
\description{
Apply a transformation on the data
}
