% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{univariate_xts}
\alias{univariate_xts}
\title{Extract a univariate xts object from a mutivariate 'xts' object}
\usage{
univariate_xts(mts_data_xts, ind = 1)
}
\arguments{
\item{mts_data_xts}{a univariate or multivariate 'xts' object}

\item{ind}{a numeric, the column index of the desired univariate 'xts' object}
}
\description{
Extract a univariate xts object from a mutivariate 'xts' object
}
