% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{valid_md_autml_h2o}
\alias{valid_md_autml_h2o}
\title{Check AutoML-h2o model validity}
\usage{
valid_md_autml_h2o(ts_data, automl_h2o_arg)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}

\item{automl_h2o_arg}{A list, optional arguments to pass to the \code{\link[h2o]{h2o.automl}} function}
}
\value{
A boolean, TRUE if the forecasting is possible.
}
\description{
This function ensures that the data meets model requirements
}
