% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{valid_md_bsts}
\alias{valid_md_bsts}
\title{Check BSTS model validity}
\usage{
valid_md_bsts(ts_data)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}
}
\value{
A boolean, TRUE if the forecasting is possible.
}
\description{
This function ensures that the data meets model requirements
}
