% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{valid_md_lstm_keras}
\alias{valid_md_lstm_keras}
\title{Check LSTM-keras model validity}
\usage{
valid_md_lstm_keras(ts_data, lstm_keras_arg)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}

\item{lstm_keras_arg}{A list, optional arguments to pass to the lstm network}
}
\value{
A boolean, TRUE if the forecasting is possible.
}
\description{
This function ensures that the data meets model requirements
}
