% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{valid_md_snaive}
\alias{valid_md_snaive}
\title{Check seasonal-naive model validity}
\usage{
valid_md_snaive(ts_data, fc_horizon)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}

\item{fc_horizon}{An integer, the forecasting horizon (i.e. the number of periods to forecast)}
}
\value{
A boolean, TRUE if the forecasting is possible.
}
\description{
This function ensures that the data meets model requirements
}
