% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_backtesting_iter}
\alias{check_backtesting_iter}
\title{Check backtesting iteration number
This function ensures that the backtesting iteration number is valid.
If not, the function throws an error.}
\usage{
check_backtesting_iter(bt_iter, backtesting_opt = NULL)
}
\arguments{
\item{bt_iter}{must be a positive integer which is smaller or equal to the
number of backtesting operations to perform.}
}
\value{
bt_iter: a positive integer
}
\description{
Check backtesting iteration number
This function ensures that the backtesting iteration number is valid.
If not, the function throws an error.
}
