% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_backtesting_opt}
\alias{check_backtesting_opt}
\title{Check backtesting options}
\usage{
check_backtesting_opt(backtesting_opt)
}
\arguments{
\item{backtesting_opt}{A list, options for the backtesting program:

 use_bt - A boolean, to determine whether to apply backtesting or to generate forcasts on future dates

 nb_iters - An integer, to determine the number of backtesting operations to apply

 method - A string, to determine whether to use a rolling or a moving forecasting window

 sample_size - A string, to determine whether the training set size should expand or remain
 fixed across backtesting operations}
}
\value{
A list of backtesting options
}
\description{
This function ensures that the selected backtesting options are valid.
If not, the function throws an error. If the options are NULL, then default
values will be applied.
}
