% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_colnames}
\alias{check_colnames}
\title{Drop punctuations in colnames}
\usage{
check_colnames(data_colnames, default_colname = "time_series", nb_colnames = 1)
}
\arguments{
\item{data_colnames}{A vector of strings, colnames of the input data which need to be cleaned}

\item{default_colname}{a string, default colname assigned when no colnames are defined}

\item{nb_colnames}{an integer, number of colnames to check}
}
\value{
data_colnames: A vector of strings, where special symbols are dropped
}
\description{
Punctuations in colnames will dropped in order to
avoid errors when converting between different data types.
}
