% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_model_names}
\alias{check_model_names}
\title{Check model names}
\usage{
check_model_names(model_names)
}
\arguments{
\item{model_names}{A list or vector of strings}
}
\value{
model_names: vector of strings
}
\description{
This function ensures that the selected model names are valid.
If not, the function throws an error.
}
