% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_models_args}
\alias{check_models_args}
\title{Check models' arguments}
\usage{
check_models_args(models_args, model_names = NULL)
}
\arguments{
\item{models_args}{A list}
}
\value{
models_args: A list
}
\description{
This function ensures that the models' arguments are a list. Furthermore, the
function checks if these arguments match the selected model names, otherwise they
will be dropped.
}
