% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_preprocess_fct}
\alias{check_preprocess_fct}
\title{Check the custom preprocessing function}
\usage{
check_preprocess_fct(prepro_fct)
}
\arguments{
\item{prepro_fct}{A function, a preprocessing function which handles missing values in the data.
The default preprocessing function selects the largest interval of non-missing values and then attributes the
most recent dates to those values. Other data handling functions can be applied (e.g. timeSeries::na.contiguous,
imputeTS::na.mean, custom-developed...).}
}
\value{
NULL (if no fct is specified) or prepro_fct (if prepro_fct is specified)
}
\description{
This function checks whether the specified function is indeed a function
}
