% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_tensorflow}
\alias{check_tensorflow}
\title{Check if tensorflow is properly installed}
\usage{
check_tensorflow(model_names)
}
\arguments{
\item{model_names}{A list or vector of strings representing the model names to be used}
}
\value{
model_names: vector of strings the input model_names without 'lstm_keras' depending on that algorithms prerequisites.
}
\description{
This function ensures that the user has Python and Tensorflow (version <= 1.14) installed.
If these criteria are violated, then the user cannot use the 'lstm_keras' model.
}
