% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_valid_set_size}
\alias{check_valid_set_size}
\title{Check validation set size
This function ensures that the selected validation set size is valid.
If not, the function throws an error.}
\usage{
check_valid_set_size(valid_set_size)
}
\arguments{
\item{valid_set_size}{must be a positive integer}
}
\value{
valid_set_size: a positive integer
}
\description{
Check validation set size
This function ensures that the selected validation set size is valid.
If not, the function throws an error.
}
