% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_checks.R
\name{check_valid_set_size}
\alias{check_valid_set_size}
\title{Check validation set size}
\usage{
check_valid_set_size(valid_set_size)
}
\arguments{
\item{valid_set_size}{An integer, the validation set size}
}
\value{
valid_set_size: An integer
}
\description{
This function ensures that the selected validation set size is valid.
If not, the function throws an error.
}
