% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{combine_fc_results}
\alias{combine_fc_results}
\title{Combine forecasting info}
\usage{
combine_fc_results(
  model_name,
  fc_formated,
  actual_formated,
  split_keys,
  model_descr = NULL,
  model_par = NULL,
  model_args = NULL
)
}
\arguments{
\item{model_name}{A string}

\item{fc_formated}{A data.frame}

\item{actual_formated}{A data.frame}

\item{split_keys}{A data.frame}

\item{model_descr}{A string}

\item{model_par}{A string}

\item{model_args}{A string}
}
\value{
A data.frame object
}
\description{
This function combines every info which will be stored in a data table.
}
