% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{format_historical_data}
\alias{format_historical_data}
\title{Format original data}
\usage{
format_historical_data(ts_data)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}
}
\value{
A data.frame object
}
\description{
This function ensures that the original (unprocessed) data input is standardized across all forecasting procedures
before saving it in a data table. The original data can be accessed under 'values' with 'key'=='actual'.
}
