% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc}
\alias{generate_fc}
\title{Forecasting Engine API}
\usage{
generate_fc(
  mts_data,
  fc_horizon = 12,
  xreg_data = NULL,
  backtesting_opt = list(use_bt = FALSE, nb_iters = 1, method = c("rolling", "moving"),
    sample_size = c("expanding", "fixed")),
  model_names = c("arima", "ets", "tbats", "bsts", "snaive", "nnetar", "stl",
    "automl_h2o"),
  preprocess_fct = NULL,
  models_args = NULL,
  save_fc_to_file = NULL,
  time_id = base::Sys.time(),
  nb_cores = 1,
  ...
)
}
\arguments{
\item{mts_data}{A univariate or multivariate ts, mts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{xreg_data}{A univariate or multivariate ts, mts or xts object, optional external regressors}

\item{backtesting_opt}{A list, options for the backtesting program:

 use_bt - A boolean, to determine whether to apply backtesting or to generate forcasts on future dates

 nb_iters - An integer, to determine the number of backtesting operations to apply

 method - A string, to determine whether to use a rolling or a moving forecasting window

 sample_size - A string, to determine whether the training set size should expand or
 remain fixed across backtesting operations}

\item{model_names}{A list or character, names of models to apply}

\item{preprocess_fct}{A function, a custom function to handle missing values in the data.
(e.g. timeSeries::na.contiguous or imputeTS::na.mean) As default the largest
interval of non-missing values is selected and the most recent dates are then attributed to these values.}

\item{models_args}{A list, optional arguments to pass to the models}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{time_id}{A POSIXct, created with \code{\link[base]{Sys.time}} and appended to results}

\item{nb_cores}{An integer, the number of threads to use for automl_h2o}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A list
}
\description{
Function which enables the user to select different forecasting algorithms ranging from
traditional time series models (i.e. ARIMA, ETS, STL) to machine learning methods (i.e. LSTM, AutoML).
}
\examples{
\dontrun{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc(AirPassengers,
                  fc_horizon = 12)

fc <- generate_fc(AirPassengers,
                  fc_horizon = 6,
                  model_names = c("arima", "ets",
                                  "lstm_keras",
                                  "automl_h2o"))
fc <- generate_fc(AirPassengers,
                  fc_horizon = 6,
                  model_names = c("ets", "snaive",
                                  "stl", "nnetar"),
                  model_args = list(ets_arg = list(model = "ZZA",
                                                   opt.crit = "amse",
                                                   upper = c(0.3, 0.2,
                                                             0.2, 0.98),
                                    stl_arg = list(s.window = "periodic"))))

# Generate forecasts on past dates to analyze performance
fc <- generate_fc(AirPassengers,
                  model_names = "arima",
                  fc_horizon = 12,
                  backtesting_opt = list(use_bt = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc(AirPassengers,
                  model_names = "tbats",
                  fc_horizon = 6,
                  backtesting_opt = list(use_bt = TRUE,
                                         nb_iters = 6))
}
}
