% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_engine_api.R
\name{generate_fc}
\alias{generate_fc}
\title{Forecasting Engine API}
\usage{
generate_fc(
  mts_data,
  fc_horizon = 12,
  xreg_data = NULL,
  backtesting_opt = list(use_bt = FALSE, nb_iters = 1, method = c("rolling", "moving"),
    sample_size = c("expanding", "fixed")),
  model_names = c("arima", "ets", "tbats", "bsts", "stl", "snaive", "nnetar"),
  models_args = NULL,
  save_fc_to_file = NULL,
  use_parallel = FALSE,
  ...
)
}
\arguments{
\item{mts_data}{A univariate or multivariate ts, mts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{backtesting_opt}{A list, options for the backtesting program
use_bt: A boolean, to determine whether to apply backtesting or to generate forcasts on future dates
nb_iters: An integer, to determine the number of backtesting operations to apply
method: A string, to determine whether to use a rolling or a moving forecasting window
sample_size: A string, to determine whether the training set size should expand or remain fixed across backtesting operations}

\item{model_names}{A list or character, names of models to apply}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{use_parallel}{A boolean, apply parallel processing}

\item{xreg_xts}{A univariate or multivariate ts, mts or xts object, optional external regressors}

\item{model_args}{A list, optional arguments to pass to the models}
}
\value{
A list, forecast object for each forecasted period
}
\description{
Function which enables the user to select different forecasting algorithms ranging from
traditional time series models (i.e. ARIMA, ETS, STL) to machine learning methods (i.e. LSTM, AutoML).
}
\examples{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc(AirPassengers,
                  fc_horizon = 12)

fc <- generate_fc(AirPassengers,
                  fc_horizon = 6,
                  model_names = c("arima", "ets",
                                  "lstm_keras",
                                  "automl_h2o"))
fc <- generate_fc(AirPassengers,
                  fc_horizon = 6,
                  model_names = c("ets", "snaive",
                                  "stl", "nnetar"),
                  model_args = list(ets_arg = list(model = "ZZA",
                                                   opt.crit = "amse",
                                                   upper = c(0.3, 0.2,
                                                             0.2, 0.98),
                                    stl_arg = list(s.window = "periodic")))
                  use_parallel = TRUE)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc(AirPassengers,
                  model_names = "arima",
                  fc_horizon = 12,
                  backtesting_opt = list(use_bt = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc(AirPassengers,
                  model_names = "tbats",
                  fc_horizon = 6,
                  backtesting_opt = list(use_bt = TRUE,
                                         nb_iters = 6))
}
