% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_arima}
\alias{generate_fc_arima}
\title{ARIMA Model:
applies the \code{\link[forecast]{auto.arima}} function from the \code{\link[forecast]{forecast}} package
on time series data.}
\usage{
generate_fc_arima(
  ts_data_xts,
  fc_horizon = 12,
  xreg_xts = NULL,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  arima_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{A univariate ts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{xreg_xts}{A univariate or multivariate ts, mts or xts object, optional external regressors}

\item{backtesting_opt}{A list, options of the backtesting program}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{arima_arg}{A list, optional arguments to pass to the \code{\link[forecast]{auto.arima}} function}
}
\value{
A list, forecast object for each forecasted period
}
\description{
ARIMA Model:
applies the \code{\link[forecast]{auto.arima}} function from the \code{\link[forecast]{forecast}} package
on time series data.
}
\examples{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc_arima(AirPassengers,
                        fc_horizon = 12)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc_arima(AirPassengers,
                        fc_horizon = 12,
                        backtesting_opt = list(use_backtesting = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc_arima(AirPassengers,
                        fc_horizon = 6,
                        backtesting_opt = list(use_backtesting = TRUE,
                                               backtesting_nb_iters = 6))
}
