% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_bsts}
\alias{generate_fc_bsts}
\title{Bayesian Structural Time Series Model}
\usage{
generate_fc_bsts(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  preprocess_fct = NULL,
  bsts_arg = NULL,
  time_id = base::Sys.time(),
  ...
)
}
\arguments{
\item{ts_data_xts}{A univariate ts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{backtesting_opt}{A list, options for the backtesting program:

 use_bt - A boolean, to determine whether to apply backtesting or to generate forcasts on future dates

 nb_iters - An integer, to determine the number of backtesting operations to apply

 method - A string, to determine whether to use a rolling or a moving forecasting window

 sample_size - A string, to determine whether the training set size should expand or
 remain fixed across backtesting operations}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{preprocess_fct}{A custom preprocessing function to deal with missing values}

\item{bsts_arg}{A list, optional arguments to pass to the \code{\link[bsts]{bsts}} function}

\item{time_id}{A POSIXct, created with \code{\link[base]{Sys.time}} and appended to results}
}
\value{
A list
}
\description{
Function to apply the \code{\link[bsts]{bsts}} function from the \code{bsts} package on
time series data.
}
\examples{
## Not run:
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc_bsts(AirPassengers,
                       fc_horizon = 12)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc_bsts(AirPassengers,
                       fc_horizon = 12,
                       backtesting_opt = list(use_bt = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc_bsts(AirPassengers,
                       fc_horizon = 6,
                       backtesting_opt = list(use_bt = TRUE,
                                              nb_iters = 6))
## End (Not run)
}
