% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_ets}
\alias{generate_fc_ets}
\title{Exponential Smoothing Model}
\usage{
generate_fc_ets(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  ets_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{A univariate ts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{backtesting_opt}{A list, options of the backtesting program}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files}

\item{ets_arg}{A list, optional arguments to pass to the \code{\link[forecast]{ets}} function}
}
\value{
A list, forecast object for each forecasted period
}
\description{
Function to apply the \code{\link[forecast]{ets}} function from the \code{\link[forecast]{forecast}} package on
time series data.
}
\examples{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc_ets(AirPassengers,
                      fc_horizon = 12)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc_ets(AirPassengers,
                      fc_horizon = 12,
                      backtesting_opt = list(use_backtesting = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc_ets(AirPassengers,
                      fc_horizon = 6,
                      backtesting_opt = list(use_backtesting = TRUE,
                                             backtesting_nb_iters = 6))
}
