% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_ets}
\alias{generate_fc_ets}
\title{Exponential Smoothing Model}
\usage{
generate_fc_ets(
  ts_data,
  fc_horizon = 12,
  backtesting_opt = NULL,
  data_dir = NULL,
  prepro_fct = NULL,
  ets_arg = NULL,
  time_id = base::Sys.time(),
  ...
)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}

\item{fc_horizon}{An integer, the forecasting horizon (i.e. the number of periods to forecast)}

\item{backtesting_opt}{A list, options which define the backtesting approach:

 use_bt - A boolean, to determine whether forecasts should be generated on future dates (default) or on past values. Generating
 forecasts on past dates allows to measure past forecast accuracy and to monitor a statistical model's ability to learn
 signals from the data.

 nb_iters - An integer, to determine the number of forecasting operations to apply (When no backtesting is selected, then only
 one forecasting exercise is performed)

 method - A string, to determine whether to apply a 'rolling' (default) or a 'moving' forecasting window. When 'rolling' is selected,
 after each forecasting exercise, the forecasting interval increments by one period and drops the last period to include it in
 the new training sample. When 'moving' is selected, the forecasting interval increments by its size rather than one period.

 sample_size - A string, to determine whether the training set size should be 'expanding' (default) or 'fixed'.
 When 'expanding' is selected, then after each forecasting operation, the periods dropped from the forecasting interval will
 be added to the training set. When 'fixed' is selected, then adding new periods to the training set will require dropping as
 many last periods to keep the set's size constant.}

\item{data_dir}{A string, directory to which results can be saved as text files}

\item{prepro_fct}{A function, a preprocessing function which handles missing values in the data.
The default preprocessing function selects the largest interval of non-missing values and then attributes the
most recent dates to those values. Other data handling functions can be applied (e.g. timeSeries::na.contiguous,
imputeTS::na.mean, custom-developped...).}

\item{ets_arg}{A list, optional arguments to pass to the \code{\link[forecast]{ets}} function}

\item{time_id}{A POSIXct, timestamp created with \code{\link[base]{Sys.time}} which is then appended to the results}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A 'tsForecastR' object
}
\description{
Function to apply the \code{\link[forecast]{ets}} function from the \code{forecast} package on
time series data.
}
\examples{
\dontrun{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc_ets(AirPassengers,
                      fc_horizon = 12)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc_ets(AirPassengers,
                      fc_horizon = 12,
                      backtesting_opt = list(use_bt = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc_ets(AirPassengers,
                      fc_horizon = 6,
                      backtesting_opt = list(use_bt = TRUE,
                                             nb_iters = 6))
}
}
