% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_lstm_keras}
\alias{generate_fc_lstm_keras}
\title{Long-Short Term Memory Network
This function applies lstm networks from the keras package on
time series data.}
\usage{
generate_fc_lstm_keras(
  ts_data_xts,
  fc_horizon = 12,
  xreg_xts = NULL,
  backtesting_opt = NULL,
  save_fc_to_file = NULL,
  lstm_keras_arg = NULL,
  ...
)
}
\arguments{
\item{ts_data_xts}{univariate xts object}
}
\value{
A list of forecast objects
}
\description{
Long-Short Term Memory Network
This function applies lstm networks from the keras package on
time series data.
}
