% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_models.R
\name{generate_fc_stl}
\alias{generate_fc_stl}
\title{Season-Trend Decomposition with Loess Model}
\usage{
generate_fc_stl(
  ts_data_xts,
  fc_horizon = 12,
  backtesting_opt = NULL,
  data_dir = NULL,
  preprocess_fct = NULL,
  stl_arg = NULL,
  time_id = base::Sys.time(),
  ...
)
}
\arguments{
\item{ts_data_xts}{A univariate ts or xts object}

\item{fc_horizon}{An integer, the forcasting horizon}

\item{backtesting_opt}{A list, options for the backtesting program:

 use_bt - A boolean, to determine whether to apply backtesting or to generate forcasts on future dates

 nb_iters - An integer, to determine the number of backtesting operations to apply

 method - A string, to determine whether to use a rolling or a moving forecasting window

 sample_size - A string, to determine whether the training set size should expand or
 remain fixed across backtesting operations}

\item{data_dir}{A string, directory to which results can be saved as text files}

\item{preprocess_fct}{A function, a custom function to handle missing values in the data.
(e.g. timeSeries::na.contiguous or imputeTS::na.mean) As default the largest
interval of non-missing values is selected and the most recent dates are then attributed to these values.}

\item{stl_arg}{A list, optional arguments to pass to the \code{\link[stats]{stl}} function}

\item{time_id}{A POSIXct, created with \code{\link[base]{Sys.time}} and appended to results}

\item{...}{Additional arguments to be passed to the function}
}
\value{
A tsForecastR object
}
\description{
Function to apply the \code{\link[stats]{stl}} function from the \code{stats} package on
time series data.
}
\examples{
\dontrun{
library(datasets)

# Generate forecasts on future dates
fc <- generate_fc_stl(AirPassengers,
                      fc_horizon = 12)

# Generate forecasts on past dates to analyze performance
fc <- generate_fc_stl(AirPassengers,
                      fc_horizon = 12,
                      backtesting_opt = list(use_bt = TRUE))

# Generate forecasts on past dates with multiple iterations and a rolling window
fc <- generate_fc_stl(AirPassengers,
                      fc_horizon = 6,
                      backtesting_opt = list(use_bt = TRUE,
                                             nb_iters = 6))
}
}
