% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_fc_with_PI}
\alias{get_fc_with_PI}
\title{Extract forecasts and prediction intervals from list}
\usage{
get_fc_with_PI(fc_obj, exclude_PI = FALSE)
}
\arguments{
\item{fc_obj}{A list, the forecasts must be stored under the keyword 'mean'.}

\item{exclude_PI}{A boolean, extract prediction intervals stored under the keywords 'upper' and 'lower'.}
}
\value{
A data.frame
}
\description{
This function extracts the forecasts and prediction intervals from a list.
}
