% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_custom_fct}
\alias{preprocess_custom_fct}
\title{Customized preprocessing function}
\usage{
preprocess_custom_fct(ts_data, transf_fct = NULL)
}
\arguments{
\item{ts_data}{A univariate 'ts' or 'xts' object}

\item{transf_fct}{A function, function which transforms the data}
}
\value{
An xts object, the transformed data
}
\description{
The user can specify a custom preprocessing function to deal with missing values
}
\examples{
\dontrun{
library(datasets)
library(timeSeries)
preprocessing(AirPassengers, timeSeries::na.contiguous)

library(imputeTS)
preprocessing(AirPassengers, imputeTS::na.mean)
}
}
