% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_custom_fct}
\alias{preprocess_custom_fct}
\title{Customized preprocessing function}
\usage{
preprocess_custom_fct(input_data, fct = NULL)
}
\arguments{
\item{ts_data_xts}{A ts, mts or xts object}
}
\value{
An xts object
}
\description{
The user can specify a custom preprocessing function to deal with missing values
}
\examples{
## Not run:
library(datasets)
library(timeSeries)
preprocessing(AirPassengers, timeSeries::na.contiguous)

library(imputeTS)
preprocessing(AirPassengers, imputeTS::na.mean)
## End (Not run)
}
