% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{read_fc_from_file}
\alias{read_fc_from_file}
\title{Read results from files}
\usage{
read_fc_from_file(data_colnames, save_fc_to_file, model_names)
}
\arguments{
\item{save_fc_to_file}{A directory}
}
\value{
A data frame
}
\description{
Read results from files
}
