% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{read_fc_from_file}
\alias{read_fc_from_file}
\title{Read results from files}
\usage{
read_fc_from_file(data_colnames, data_dir, model_names)
}
\arguments{
\item{data_colnames}{A vector of strings, colnames of the input data which need to be cleaned}

\item{data_dir}{A string, directory to which results can be saved as text files}

\item{model_names}{A list or vector of strings representing the model names to be used}
}
\value{
A data frame
}
\description{
Read results from files
}
