% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{save_as_df}
\alias{save_as_df}
\title{Read the results}
\usage{
save_as_df(
  fc = NULL,
  save_fc_to_file = NULL,
  data_colnames = NULL,
  model_names = NULL,
  ...
)
}
\arguments{
\item{fc}{A tsForecastR object}

\item{save_fc_to_file}{A string, directory to which results can be saved as text files.}

\item{data_colnames}{A vector of strings, the names of the time series objects to read the
results from}

\item{model_names}{A vector of strings, the models to read the results from}
}
\value{
A data frame
}
\description{
This function transforms the tsForecastR object into a data.frame object
}
