% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{split_train_test_set}
\alias{split_train_test_set}
\title{Split the data into train, validation and test set}
\usage{
split_train_test_set(
  input_data,
  fc_horizon = 12,
  bt_iter = 1,
  valid_set_size = 0,
  tmp_test_set_size = 0,
  backtesting_opt = NULL,
  ...
)
}
\value{

}
\description{
Split the data into train, validation and test set
}
