% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{split_train_test_set}
\alias{split_train_test_set}
\title{Split the data into a training, validation and test set}
\usage{
split_train_test_set(
  input_data,
  fc_horizon = 12,
  bt_iter = 1,
  valid_set_size = 0,
  tmp_test_set_size = 0,
  backtesting_opt = NULL,
  ...
)
}
\arguments{
\item{input_data}{A univariate or multivariate ts, mts or xts object}

\item{fc_horizon}{An integer, forecasting horizon}

\item{bt_iter}{An integer, number of the current backtesting operation}

\item{valid_set_size}{An integer, the validation set size (default = 0)}

\item{tmp_test_set_size}{An integer, size of a second test set (used by h2o.automl) (default = 0)}

\item{backtesting_opt}{A list, options for the backtesting program
use_bt: A boolean, to determine whether to apply backtesting or to generate forcasts on future dates
nb_iters: An integer, to determine the number of backtesting operations to apply
method: A string, to determine whether to use a rolling or a moving forecasting window
sample_size: A string, to determine whether the training set size should expand or remain fixed across backtesting operations}
}
\value{
A list, training, validation and test sets
}
\description{
Split the data into a training, validation and test set
}
