% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transform_data}
\alias{transform_data}
\title{Apply a transformation on the data}
\usage{
transform_data(
  original_ts,
  transformed_ts = NULL,
  method = "diff",
  apply_transform = TRUE
)
}
\description{
Apply a transformation on the data
}
